Build simple game where you can drive around the screen
Implemented car classes init, update, and rotate, which collectively
allow for the initiation of a car object which will move around the
screen in the direction it is pointing, and can be steered by calling
the rotate method. Also added a few different colored car images.
Started basic game function to play the game and update over time (also
tracks time appropriately) and started working on the map class which
has methods to read an image from the webcam and covert it to a binary
image. Ultimately this image will code the map, but that has not been
implemented yet. Note, added the PIL library for map image stuff, might
need to download that to run the code.