Complete map reading in and make car interact with it
Implemented build method in the map class which will take a binary .png
file and build out the road attribute of the map class, which is a
matrix the size of the screen where each element represents weather the
corresponding pixel is part of the road. Added functionality to game to
ask a user if they wish to capture a new map at the start, and then
allows them to pick a map from a list of available image files. If the
binary version of the image doesn't already exist it will then build it
for them. Updated car class to check for collisions with non road each
update and react accordingly. Added some folders to store map images,
and a few maps I have been using for testing, one created via image
capture and the other drawn in illustrator. Added comments to code thus
far.